Add agent launch code, runtime, profile and container controllers to Simulation
Fix Location and integrate ClientsDesc and TechniciansDesc with ClientType and TechnicianType

+++ src/utils/Location.java
@@ -3,11 +3,10 @@ package utils;
 import java.io.Serializable;
 
 public class Location implements Serializable {
-
     private final double x;
     private final double y;
 
-    public Location(int x, int y) {
+    public Location(double x, double y) {
         this.x = x;
         this.y = y;
     }
